Don't select the base terminal profile by name.
The script obtained the default profile based on its name. This is brittle and
won't work on non-english locales. The change obtains the profile UUID by
querying the appropiate dconf key.

As a side effect, the terminal_source_profile option is no longer available, so
it is no longer possible to specify a different base profile for the project's
terminal (shouldn't be a big issue, but we may bring this back if there's
demand for it in the future).

Fixes #1