Mercurial should be disabled by default.
Since mercurial was enabled by default, and the --mercurial flag only allows
to *enable* it, it was impossible to have a project without mercurial enabled.

The default has been changed to False, so omitting the flag will produce a
project without a mercurial repository. To create the repository it is now
required to supply the flag.