Fixed permission issues when installing bundles and/or using --dedicated-user
Mercurial has trouble running under seteuid. Introduced an _hg() method to run
the mercurial binary using a forked process and setuid() instead. Also, the
order in which bundles are extracted and installed has changed so that the
bundle contents belong to the project's dedicated user.

As a side effect, some woostproject3 behavior has been back ported to
woostproject2:

- When cloning a required Mercurial repository, if a copy of the project already
  existed at the target location, pull and update the working copy, instead of
  doing nothing

- Added the --revision parameter

- setup_python_package() should delegate to Installer._exec

- Pass --update to rsync when copying the database file